Moved plotting to velGraph so u can decide between rot/trans just once.

+++ src/org/usfirst/frc/team1197/TorTrajectoryLib/TorTrajectory.java
@@ -267,6 +267,6 @@ public abstract class TorTrajectory {
 		return Math.signum(goal_pos)*max_jeta;
 	}
 	public double totalTime(){
-		return ((double)time.get(time.size()-1))*0.001;
+		return (time.get(time.size()-1))*0.001;
 	}
 }
